emacs: Prefer CIDER's completion to lsp-mode
For clojure-mode, at the very least, there is a bug in lsp-mode's
company completions where, when you have a completion list active
and you hit backspace to delete characters you've typed, if you
then go on to retype those same characters, lsp-mode will no longer
trigger a company completion. This is a bug we should investigate
later on.

Also do not start 'impatient-mode' by default for HTTP
file buffers since this requires starting another server
process (which is 'simple-httpd'). We should choose when and
if we want to start impatient-mode on an as-needed basis.