emacs: Disable CIDER dynamic font feature for now
There appears to be a problem with CIDER's
`cider-font-lock-dynamically` custom variable. It used to work fairly
well with `clojure-ts-mode` but is now disabling font locking in Clojure
buffers altogether. New Clojure buffers open with no syntax highlighting,
then slowly receive bits of colorization as the user edits parts of the
buffer. This is obviously not acceptable, and I will open an issue in
the CIDER GitHub project to resolve this problem.